+++ README.md
@@ -5,6 +5,8 @@ A quickie little utility which makes it easier to write tests using [specs2](htt
 ## Specs2
 
 ```scala
+import cats.effect.IO
+import cats.effect.testing.specs2.CatsIO
 import org.specs2.mutable.Specification
 
 // for some reason, only class works here; object will not be detected by sbt
