+++ specs2/src/main/scala/cats/effect/testing/specs2/CatsIO.scala
@@ -14,23 +14,15 @@
  * limitations under the License.
  */
 
-package cats.effect.specs2
+package cats.effect.testing.specs2
 
-import cats.effect.Effect
-import cats.effect.syntax.effect._
-import org.specs2.execute.{AsResult, Failure, Result}
+import cats.effect.{ContextShift, IO, Timer}
 
-import scala.concurrent.duration._
-import scala.language.higherKinds
+import scala.concurrent.ExecutionContext
 
-trait CatsEffect {
+trait CatsIO extends CatsEffect {
+  private val executionContext: ExecutionContext = ExecutionContext.global
 
-  protected val Timeout: Duration = 10.seconds
-
-  implicit def effectAsResult[F[_]: Effect, R](implicit R: AsResult[R]): AsResult[F[R]] = new AsResult[F[R]] {
-    def asResult(t: => F[R]): Result =
-      t.toIO.unsafeRunTimed(Timeout)
-        .map(R.asResult(_))
-        .getOrElse(Failure(s"expectation timed out after $Timeout"))
-  }
+  implicit val ioContextShift: ContextShift[IO] = IO.contextShift(executionContext)
+  implicit val ioTimer: Timer[IO] = IO.timer(executionContext)
 }
